% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomerate.R
\name{agglomerate}
\alias{agglomerate}
\title{agglomerates hierarchically nested codes (ICD10, ATC, etc.) to fewer, but more widespread categories than the initial input}
\usage{
agglomerate(
  data,
  domain,
  id = "pnr_enc",
  selection = NULL,
  skip = 0,
  levels,
  threshold = 0.01,
  censor.time = NULL,
  selection.pos = 1,
  free = F,
  fixed = NULL
)
}
\description{
agglomerates hierarchically nested codes (ICD10, ATC, etc.) to fewer, but more widespread categories than the initial input
}
