% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbs.R
\name{charl_score}
\alias{charl_score}
\title{charlson index scoring function}
\usage{
charl_score(x, cats = F)
}
\arguments{
\item{x}{is a data frame with columns corresponding to the names defined in comorbs$charl_names}

\item{cats}{is a logical indication of whether output should be the categorical 0, 1, 2 score or just the default count variable

@return returns a vector of length = nrow(x) with charlson scores

@export}
}
\description{
function takes a data frame with Charlson diagnosis groups as columns and generates a Charlson comorbidity index score for each row
}
