% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{form.fix}
\alias{form.fix}
\title{formats a numeric vector to a fixed string format with digits}
\usage{
form.fix(x, digits = 3, percent = F)
}
\description{
formats a numeric vector to a fixed string format with digits
}
