% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grit.R
\docType{package}
\name{grit}
\alias{grit}
\alias{grit-package}
\title{grit}
\description{
A package designed to document the analysis of repeatedly measured grit scores on the same subjects in different contexts
}

