% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.pull.R
\name{local.pull}
\alias{local.pull}
\title{pulling data from local RDS database}
\usage{
local.pull(path, vw, ids = NULL, vars = NULL, id.type = "pnr")
}
\arguments{
\item{path}{path to local project-specific database (db_build output folder)}

\item{vw}{name of view file / SQL table to pull}

\item{ids}{optional vector of ids to pull (ignoring all non-matching rows). If ids are record ids, specify id.type as 'rid'.}

\item{vars}{optional list of variables to pull.}

\item{id.type}{'pnr' for personal id number, 'pid' for using record ids as key.}
}
\value{
returns an RDS file containing specified source data. If ids are specified, only rows matching an id are returned.
}
\description{
pulling data from local RDS database
}
