% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.count.R
\name{observe.count}
\alias{observe.count}
\title{count observations from one 'value' variable containing qualitative codes of any kind}
\usage{
observe.count(data, id, index.date, value, interval, valid.values, bin = F)
}
\arguments{
\item{data}{is a dataframe with an id column, an index.date column, a value column and a date column ('date')}

\item{id}{name of id column (multiple instances within id are aggregated to one row)}

\item{index.date}{name of column containing and index/reference date}

\item{value}{name of column with values to be counted and aggregated by id}

\item{interval}{two outer limits as days (integer) relative to index. Positive integers look ahead, negative integers look back, (-10, +10) will observe 10 days around (pre/post) the index date}
}
\value{
returns a data frame of one row per id, that counts all incidences of all valid values within interval of days relative to index date
}
\description{
count observations from one 'value' variable containing qualitative codes of any kind
}
