% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.incid.R
\name{observe.incid}
\alias{observe.incid}
\title{observe date of incidence for any value (qualitative code, e.g. ICD-10 codes or derivates)}
\usage{
observe.incid(data, id, index.date, value, valid.values, relative = F)
}
\arguments{
\item{data}{a data frame with a 'date' column, an index.date column, an id column, a value column to be reduced to one row per id with the timestamp of the earliest observation of any valid value.}

\item{id}{name of id column.}

\item{index.date}{name of index date column; used for relative measures of incidence}

\item{value}{name of column with actual value corresponding to the label of metric contained in the code column}

\item{valid.values}{optional list of valid codes to weed out irrelevant measures}

\item{relative}{logical indicating whether or not incidence dates are number of days relative to index.date}
}
\value{
a data frame with one row per id and columns for each valid type of value containing the timestamp (date) of incidence
}
\description{
a long dataset (repeated measures within id) with a set of timestamped qualitative codes will be reduced to just the date of first occurence each code before the deadline (index date)
}
