% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.last.R
\name{observe.last}
\alias{observe.last}
\title{observe the last numeric value observed for any id in any interval}
\usage{
observe.last(
  data,
  id,
  index.date,
  code,
  value,
  interval,
  valid.codes,
  text = F
)
}
\arguments{
\item{data}{a data frame with a 'date' column, an index.date column, an id column, a code (metric id) column and a value column to be reduced to one row per id with the latest observation of any valid code.}

\item{id}{name of id column.}

\item{index.date}{name of index date column}

\item{code}{name of column containing the metric id (when several metrics are just stacked in a long dataframe)}

\item{value}{name of column with actual value corresponding to the label of metric contained in the code column}

\item{valid.codes}{optional list of valid codes to weed out irrelevant measures}

\item{text}{a boolean indicator of whether values are categorical / contains letters.}

\item{intervel}{the limits of an interval expressed as integer days relative to the index date}
}
\value{
a data frame with one row per id and columns with counts of input vars within interval
}
\description{
a long dataset (repeated measures within id) with a qualitative code (metric id) and its coresponding value will be reduced to just the last value of any set of codes before the deadline (index date)
}
