% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.wide.R
\name{observe.wide}
\alias{observe.wide}
\title{count observations from several columns in some interval}
\usage{
observe.wide(data, id, index.date, vars, interval)
}
\arguments{
\item{data}{a data frame with a 'date' column, an index.date column, an id column and a set of count variables to be reduced to one row per id (n = length(unique(ids))).}

\item{id}{name of id column.}

\item{index.date}{name of index date column}

\item{vars}{vector of names of the count variables to be reduced}

\item{intervel}{the limits of an interval expressed as integer days relative to the index date}
}
\value{
a data frame with one row per id and columns with counts of input vars within interval
}
\description{
reduces repeated measures (for id) of a set of count varaibles
}
