% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation.R
\name{tab.it}
\alias{tab.it}
\title{wraps up (align & stack)  a set of twoway tables}
\usage{
tab.it(
  data,
  case,
  xs,
  ctns,
  bins = NULL,
  cats = NULL,
  low.cens = 10,
  decimals.p = 1,
  decimals.sd = 1
)
}
\description{
wraps up (align & stack)  a set of twoway tables
}
