% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation.R
\name{tabulate}
\alias{tabulate}
\title{wraps up (align & stack)  a set of twoway tables}
\usage{
tabulate(
  data,
  xs,
  treat,
  num = NA,
  cat = NA,
  bin = NA,
  dichotomize = NA,
  cal.date = NA,
  cens = 5,
  show.na = F
)
}
\description{
wraps up (align & stack)  a set of twoway tables
}
