% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation.R
\name{twoway.num}
\alias{twoway.num}
\title{twoway table of numeric data}
\usage{
twoway.num(
  data,
  x,
  group,
  digit.m = 1,
  digit.sd = 1,
  inf = FALSE,
  cal.date = F
)
}
\description{
twoway table of numeric data
}
