% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{view.it}
\alias{view.it}
\title{Inspection of big dictionaries}
\usage{
view.it(x, col, key, illegal = NULL, subs = NULL)
}
\value{
returns the 'rows of 'col' that are identified using the 'key' and the optional 'illegal' keywords.
}
\description{
Inspection of big dictionaries
}
